+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refund modal blocked on loading
 - Apple Pay transactions not showing up on order detail page
 - Fixed a fatal error caused by return types in `SettingsLoader::deserialize()` and `MainSettingsUpdater::update()`
+- Fixed order status being downgraded by a notification
 
 ## [3.4.1] - 2026-08-10